refactor: 필요 없는 테스트 제거
