+++ src/test/java/domain/line/NonContinuousLineGeneratorTest.java
@@ -1,6 +1,6 @@
-package domain;
+package domain.line;
 
-import domain.line.NonContinuousLineGenerator;
+import domain.ConnectionStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
