+++ src/test/java/domain/name/NameTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.name;
 
-import domain.name.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
