+++ src/test/java/domain/name/NamesTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.name;
 
-import domain.name.Names;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
